Fix Coined.set_marked()
Set marked was not working on grids if vertices were represented using coordinates. In addition, if a single vertex was passed (outside of a list), it would be inserted automatically in a list. This situation is now forbidden. Otherwise, it would be impossible to determine whether (0, 1) referred to vertices with numbers 0 and 1 or to the vertex with coordinates (0, 1).