Remove generator
After some discussions, we concluded that the quantum_walk_generator function was an overcomplicated interface for generator expressions. We expect users to be familiar with using generator expressions for list comprehensions. Mentioning in the documentation that it is recommended to use a generator instead of a list in certain places and explaining how to do it will suffice.