Update ContinuousTime.simulate
The previous behavior was confusing. The simulate method updated the evolution operator, which is beyond the simulation's responsabilities. To tackle this problem, simulate() was changed to handle integer tuples only. Float values of time can be specified via the set_time() or set_evolution() methods. The simulatte() method is responsible for applying the stored evolution operator a discrete number of times to the initial state. The simulation results are states corresponding to evenly spaced timestamps